requirements.txt
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
